package refactor from mil.nga.giat.geopackage.mapcache to mil.nga.mapcache
