+++ mapcache/build.gradle
@@ -8,7 +8,7 @@ android {
     buildToolsVersion "21.1.2"
 
     defaultConfig {
-        applicationId "mil.nga.giat.geopackage.mapcache"
+        applicationId "mil.nga.mapcache"
         minSdkVersion 17
         targetSdkVersion 21
     }
@@ -44,7 +44,7 @@ android {
 }
 
 dependencies {
-    compile "mil.nga.giat.geopackage.android:geopackage-sdk:1.0.0"
+    compile "mil.nga.geopackage.android:geopackage-sdk:1.0.0" // comment out to build locally
     //compile project(':geopackage-sdk') // uncomment me to build locally
     compile 'com.google.android.gms:play-services:6.1.71'
 }
