+++ mapcache/src/main/java/mil/nga/mapcache/data/GeoPackageTable.java
@@ -1,4 +1,4 @@
-package mil.nga.giat.geopackage.mapcache.data;
+package mil.nga.mapcache.data;
 
 import java.io.Serializable;
