+++ mapcache/src/main/res/layout/fragment_navigation_drawer.xml
@@ -6,4 +6,4 @@
     android:choiceMode="singleChoice"
     android:divider="@android:color/transparent"
     android:dividerHeight="0dp"
-    tools:context="mil.nga.giat.geopackage.mapcache.NavigationDrawerFragment" />
+    tools:context="mil.nga.mapcache.NavigationDrawerFragment" />
