+++ mapcache/src/main/res/menu/main.xml
@@ -1,6 +1,6 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="mil.nga.giat.geopackage.mapcache.MainActivity" >
+    tools:context="mil.nga.mapcache.MainActivity" >
 
     <group android:id="@+id/menu_group_map" >
         <item
